QueryPlaceCandidate 수정
todo
완전히 없을 때의 예외처리를 해야함

+++ Pang/Assets/Scripts/Item/IItemContainer.cs
@@ -31,7 +31,7 @@ public class ItemStack
 	private float itemSize => itemDB.GetItemSize(itemID);
 	public float AvailableSpace => (maxStackSize - (itemSize * quantity));
 
-	public int AvailableAmount => (int)AvailableSpace / (int)itemSize;
+	public int AvailableAmount => (int)(AvailableSpace / itemSize);
 
 	public uint ItemID => itemID;
 	public int Quantity => quantity;
